% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{cat_functionn}
\alias{cat_functionn}
\title{A Cat Function}
\usage{
cat_functionn(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of catssssssss and birds.
}
\examples{
cat_function()
}
\keyword{cats}
